+++ NMSIS/Core/Include/nmsis_core.h
@@ -99,6 +99,8 @@
 #include "core_feature_smpcc.h"
 /* Include core ecc feature header file */
 #include "core_feature_ecc.h"
+/* Include core iregion info header file */
+#include "core_feature_iinfo.h"
 /* Include compatiable functions header file */
 #include "core_compatiable.h"